Create README.md
There are several, models attempting to forecast the failure of operation of the treatment system. One such model, I contributed in working in
is the LSTM neural network model. Feature engineering was applied, and the input data set was normalized. The main function imports the model, into an interactive dashboard, webpage dashboard, via plot.ly. 
The data is unfortunately private and is no longer accessible.